fix(react-billing): resolve lint errors in SubscriptionPanel stories
The main stories file exceeded the 400-line max-lines rule. Split
AnnualSubscription, ClickableMetrics, and ManyActions into a separate
SubscriptionPanelManyActions.stories.tsx file and fixed arrow-function
style to satisfy prefer-arrow-functions/prefer-arrow-functions.
